+++ changelog.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 ## V1.15.03 - 2020-10-11
 - Minor update to correct issue with writing config file
 - Added additional checks to look for corrupt config file and restore the file if corrupted
+- Changed some add on fields to hide passwords
+- Allow external tank data to be used in calculating time remaining until tank empty
 
 ## V1.15.02 - 2020-10-07
 - Minor fix to allow some pre-Nexus models to correctly perform a model lookup