+++ changelog.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Allow external tank data to be used in calculating time remaining until tank empty
 - Update typos
 - Update to gensnmp.py to fix issue with H100 alarm list
+- Update to include fuel remainging based on estimated and current load assumptions
 
 ## V1.15.02 - 2020-10-07
 - Minor fix to allow some pre-Nexus models to correctly perform a model lookup